Render errors received from inventory API
The frontend handles errors of type KubeClientError that are returned
from the K8s API properly. However, errors that are returned from the
inventory API are of a different type and it's not trivial to tweak
gin-gonic to return such errors.

Therefore, changing the frontend to render errors that are returned from
the inventory API differently, these errors are identified by the header
'forkift-error-message'.

Signed-off-by: Arik Hadas <ahadas@redhat.com>